+++ tests/test_format.py
@@ -13,27 +13,64 @@ from inifix.format import main
 @pytest.mark.parametrize("flag", ["-i", "--inplace"])
 def test_format_keep_data(flag, inifile, capsys, tmp_path):
     target = tmp_path / inifile.name
-    data_ref = load(inifile)
+
+    ref_data = load(inifile)
     shutil.copyfile(inifile, target)
 
     ret = main([str(target), flag])
-    assert ret == 0
+    assert isinstance(ret, int)
+
     out, err = capsys.readouterr()
+
+    # nothing to output to stdout with --inplace
     assert out == ""
+    if err == f"{target} is already formatted\n":
+        assert ret == 0
+    else:
+        assert err == f"Fixing {target}\n"
+        assert ret != 0
 
     data_new = load(target)
-    assert data_new == data_ref
+    assert data_new == ref_data
+
+
+@pytest.mark.parametrize("infile", ("format-in.ini", "format-out.ini"))
+def test_exact_format_stdout(infile, capsys, tmp_path):
+    DATA_DIR = Path(__file__).parent / "data"
+
+    expected = (DATA_DIR / "format-out.ini").read_text() + "\n"
+
+    target = tmp_path / "out.ini"
+    shutil.copyfile(DATA_DIR / infile, target)
+    baseline = target.read_text() + "\n"
+
+    ret = main([str(target)])
+    out, err = capsys.readouterr()
+
+    assert out == expected
+    if err == f"{target} is already formatted\n":
+        assert ret == 0
+        assert out == baseline
+    else:
+        assert err == f"Fixing {target}\n"
+        assert ret != 0
+        assert out != ""
 
 
 @pytest.mark.parametrize("flag", ["-i", "--inplace"])
-def test_exact_format(flag, capsys, tmp_path):
+def test_exact_format_inplace(flag, capsys, tmp_path):
     DATA_DIR = Path(__file__).parent / "data"
     target = tmp_path / "out.ini"
     shutil.copyfile(DATA_DIR / "format-in.ini", target)
+
     ret = main([str(target), flag])
-    assert ret == 0
     out, err = capsys.readouterr()
-    assert out == ""
+
+    if err == f"{target} is already formatted\n":
+        assert ret == 0
+    else:
+        assert err == f"Fixing {target}\n"
+        assert ret != 0
 
     expected = (DATA_DIR / "format-out.ini").read_text()
     res = target.read_text()
@@ -46,10 +83,12 @@ def test_exact_format_with_column_size_flag(size, capsys, tmp_path):
     DATA_DIR = Path(__file__).parent / "data"
     target = tmp_path / "out.ini"
     shutil.copyfile(DATA_DIR / "format-column-size-in.ini", target)
+
     ret = main([str(target), "-i", "--name-column-size", size])
-    assert ret == 0
     out, err = capsys.readouterr()
-    assert out == ""
+
+    assert err == f"Fixing {target}\n"
+    assert ret != 0
 
     expected = (DATA_DIR / f"format-column-size-out-{size}.ini").read_text()
     res = target.read_text()
@@ -71,7 +110,7 @@ def test_empty_file(capsys, tmp_path):
     ret = main([str(target)])
     assert ret != 0
     out, err = capsys.readouterr()
-    assert out == ""
+    assert out == "\n"
     assert f"Error: {target} appears to be emtpy.\n" in err
 
 
@@ -96,10 +135,17 @@ def test_write_to_console(inifile, capsys, tmp_path):
     target = tmp_path / inifile.name
     shutil.copy(inifile, target)
 
-    # format the file for easier comparison
-    ret = main([str(target), "--inplace"])
+    baseline = target.read_text() + "\n"
 
     ret = main([str(target)])
-    assert ret == 0
+    # we can't predict if formatting is needed
+    assert isinstance(ret, int)
     out, err = capsys.readouterr()
-    assert out == ""
+
+    if err == f"{target} is already formatted\n":
+        assert ret == 0
+        assert out == baseline
+    else:
+        assert err == f"Fixing {target}\n"
+        assert ret != 0
+        assert out != ""