+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.10.0] - 2021-11-23
 
-ENH: expose name column length parameter to users in inifix-format
+- ENH: expose name column length parameter to users in inifix-format
 [PR #73](https://github.com/neutrinoceros/inifix/pull/73)
+- BUG: fix formatter behabiour
+[PR #83](https://github.com/neutrinoceros/inifix/pull/83)
 
 ## [0.9.1] - 2021-11-21